v0.9.3.5 GELU activation, optimizations for faster inference to support more parallel agents on same performance, changed training approach for faster convergence and time, other minor fixes and tweaks

+++ Assets/DeepUnity/ReinforcementLearning/Buffers/SensorBuffer.cs
@@ -71,7 +71,7 @@ namespace DeepUnity
         /// <param name="observation"></param>
         public void AddObservation(int observation)
         {
-            AddObservation(observation);
+            AddObservation((float)observation);
         }
         /// <summary>
         /// Adds an observation vector of length 2.
